feat: add debug rpc interface
